requirements.txt
torch
scipy
scikit-learn
numpy
pandas
matplotlib
pillow
tqdm
jupyter
ftfy
regex
git+https://github.com/openai/CLIP.git
faiss-cpu==1.6.5
pytorch-metric-learning
webcolors
tabulate
ipympl
lxml
brokenaxes